added empty html-part to make_payment_transaction_email and moved edit_email_html_part_and_recipients to mailing utils
